requirements.txt
Django==4.2.3
django-environ==0.10.0
ipdb==0.13.13
flake8==6.0.0
pytest==7.4.0
pytest-django==4.5.2
faker==19.2.0
django-widget-tweaks==1.4.12
django-taggit==4.0.0
pytest-cov==4.1.0
markdown==3.4.4
psycopg2-binary==2.9.7
pytest-factoryboy==2.5.1
celery[redis]==5.3.1
daphne==4.0.0